uri_aplogger/requirements.txt
asttokens==3.0.0
comm==0.2.3
contourpy==1.3.0
cycler==0.12.1
Cython==3.1.6
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
fonttools==4.60.1
future==1.0.0
h5py==3.14.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.31.0
ipython==8.18.1
iso8601==2.1.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
matplotlib==3.9.4
matplotlib-inline==0.2.1
nest_asyncio==1.6.0
numpy==2.0.2
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
pkg_resources==0.0.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
pyudev==0.24.4
pyusb==1.3.1
PyYAML==6.0.3
pyzmq==27.1.0
seabreeze==2.11.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.2.14
zipp==3.23.0
